hook_functions: dep_add_modules: Use heuristics for a generic framebuffer
With MODULES=dep, we don't want to include the full graphics drivers
like i915 and amdgpu on systems that provide a simpler framebuffer that
can be used for a video console. The code that handles this relies on
the simultaneous existence of device directories for both the generic
and full drivers, but this is no longer the case with newer kernels and
this strategy no longer works to detect an unloaded graphics driver.

Instead, use some heuristics to decide if we should skip walking full
graphics drivers, under some assumptions:

- Assume we always have a working platform graphics driver on x86.
- Assume EFI framebuffer works if we have EFI variable support.
- Assume any simple-framebuffer device listed in the device-tree works.

These are not necessarily correct, but the exceptions aren't really
worth worrying about.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>